Made charging "safer", Fixed VDPM.
-Sets VDPM to 4.76V from 4.4V always. VDPM stand for Input based Voltage
Dyanmic Power Management and monitors the incoming (USB) voltage to make
sure its not drawing too much current. You could call this charging source
protection. This change will make it more tolerant towards chargers that
implement their own "protection" when voltage drops and shuts off. This
change may become an issue if the charger can't output 4.76V or higher
because the phone simply won't charge. If this is the case your charger
sucks and you need to replace it.

Signed-off-by: ic3man5 <ic3man5@gmail.com>